+++ crates/axum-otel/src/on_failure.rs
@@ -6,7 +6,8 @@ use tracing_otel_extra::dyn_event;
 ///
 /// Original implementation from [tower-http](https://github.com/tower-rs/tower-http/blob/main/tower-http/src/trace/on_failure.rs).
 ///
-/// This component updates the span's `otel.status_code` to "ERROR" when a server error occurs.
+/// This component updates the span's `otel.status_code` and `otel.status_description`
+/// when a server error occurs.
 ///
 /// # Example
 ///
@@ -63,6 +64,10 @@ impl OnFailure<ServerErrorsFailureClass> for AxumOtelOnFailure {
         match failure_classification {
             ServerErrorsFailureClass::StatusCode(status) if status.is_server_error() => {
                 span.record("otel.status_code", "ERROR");
+                span.record(
+                    "otel.status_description",
+                    tracing::field::display(failure_classification),
+                );
             }
             _ => {}
         }